# Optimize tree size computation and the scene tree dock filter